added FAB on trip and category, added input field on trip and category, pending finalizaiton of input dates and icon selection
