+++ app/src/main/AndroidManifest.xml
@@ -11,14 +11,27 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Gabay"
-
         tools:targetApi="31">
+
         <activity
-            android:name=".LoginActivity"
-            android:exported="false" />
+            android:name=".ui.categories.CreateCategoryActivity"
+            android:exported="false"
+            android:label="Create Category"/>
+
         <activity
-            android:name=".RegisterActivity"
-            android:exported="true">
+            android:name=".ui.trips.CreateTripActivity"
+            android:exported="false"
+            android:label="Create Trip"/>
+
+        <activity
+            android:name=".ui.auth.LoginActivity"
+            android:exported="false"
+            android:theme="@style/Theme.Auth" />
+
+        <activity
+            android:name=".ui.auth.RegisterActivity"
+            android:exported="true"
+            android:theme="@style/Theme.Auth">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -28,7 +41,7 @@
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:theme="@style/Theme.MainActivity"/>
+            android:theme="@style/Theme.Gabay" />
     </application>
 
 </manifest>
