--- app/src/main/java/com/digiview/gabay/TripsFragment.java
@@ -1,22 +1,0 @@
-package com.digiview.gabay;
-
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-public class TripsFragment extends Fragment {
-
-    public TripsFragment(){
-        // require a empty public constructor
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_trips, container, false);
-    }
-}
