+++ app/src/main/java/com/digiview/gabay/ui/categories/CategoriesFragment.java
@@ -1,0 +1,49 @@
+package com.digiview.gabay.ui.categories;
+
+import android.content.Intent;
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.digiview.gabay.R;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+
+public class CategoriesFragment extends Fragment {
+
+    public CategoriesFragment(){
+        // require a empty public constructor
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_categories, container, false);
+    }
+
+    @Override
+    public void onResume() {
+        super.onResume();
+        FloatingActionButton fab = getActivity().findViewById(R.id.main_fab);
+        fab.show(); // Make the FAB visible
+
+        fab.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                // Handle FAB click for TripsFragment
+                 Intent intent = new Intent(getActivity(), CreateCategoryActivity.class);
+                 startActivity(intent);
+            }
+        });
+    }
+
+    @Override
+    public void onPause() {
+        super.onPause();
+        FloatingActionButton fab = getActivity().findViewById(R.id.main_fab);
+        fab.hide(); // Hide the FAB when the fragment is not visible
+    }
+}
