+++ app/src/main/res/drawable/ic_bus_solid.xml
@@ -1,5 +1,9 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="21dp" android:viewportHeight="512" android:viewportWidth="576" android:width="23.625dp">
-      
-    <path android:fillColor="#FF000000" android:pathData="M288,0C422.4,0 512,35.2 512,80L512,96l0,32c17.7,0 32,14.3 32,32v64c0,17.7 -14.3,32 -32,32l0,160c0,17.7 -14.3,32 -32,32v32c0,17.7 -14.3,32 -32,32L416,512c-17.7,0 -32,-14.3 -32,-32L384,448L192,448v32c0,17.7 -14.3,32 -32,32L128,512c-17.7,0 -32,-14.3 -32,-32l0,-32c-17.7,0 -32,-14.3 -32,-32l0,-160c-17.7,0 -32,-14.3 -32,-32L32,160c0,-17.7 14.3,-32 32,-32h0L64,96h0L64,80C64,35.2 153.6,0 288,0zM128,160v96c0,17.7 14.3,32 32,32L272,288L272,128L160,128c-17.7,0 -32,14.3 -32,32zM304,288L416,288c17.7,0 32,-14.3 32,-32L448,160c0,-17.7 -14.3,-32 -32,-32L304,128L304,288zM144,400a32,32 0,1 0,0 -64,32 32,0 1,0 0,64zM432,400a32,32 0,1 0,0 -64,32 32,0 1,0 0,64zM384,80c0,-8.8 -7.2,-16 -16,-16L208,64c-8.8,0 -16,7.2 -16,16s7.2,16 16,16L368,96c8.8,0 16,-7.2 16,-16z"/>
-    
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="576"
+    android:viewportHeight="512">
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M288,0C422.4,0 512,35.2 512,80L512,96l0,32c17.7,0 32,14.3 32,32v64c0,17.7 -14.3,32 -32,32l0,160c0,17.7 -14.3,32 -32,32v32c0,17.7 -14.3,32 -32,32L416,512c-17.7,0 -32,-14.3 -32,-32L384,448L192,448v32c0,17.7 -14.3,32 -32,32L128,512c-17.7,0 -32,-14.3 -32,-32l0,-32c-17.7,0 -32,-14.3 -32,-32l0,-160c-17.7,0 -32,-14.3 -32,-32L32,160c0,-17.7 14.3,-32 32,-32h0L64,96h0L64,80C64,35.2 153.6,0 288,0zM128,160v96c0,17.7 14.3,32 32,32L272,288L272,128L160,128c-17.7,0 -32,14.3 -32,32zM304,288L416,288c17.7,0 32,-14.3 32,-32L448,160c0,-17.7 -14.3,-32 -32,-32L304,128L304,288zM144,400a32,32 0,1 0,0 -64,32 32,0 1,0 0,64zM432,400a32,32 0,1 0,0 -64,32 32,0 1,0 0,64zM384,80c0,-8.8 -7.2,-16 -16,-16L208,64c-8.8,0 -16,7.2 -16,16s7.2,16 16,16L368,96c8.8,0 16,-7.2 16,-16z"/>
 </vector>
