+++ app/src/main/res/drawable/input_container.xml
@@ -1,0 +1,5 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="#ffffff"/> <!-- Change the color as needed -->
+    <corners android:radius="12dp"/> <!-- Adjust the radius to change the roundness -->
+</shape>
