+++ app/src/main/res/drawable/input_underline_state.xml
@@ -1,0 +1,4 @@
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_focused="true" android:drawable="@drawable/input_underline_focused" />
+    <item android:drawable="@drawable/input_underline_unfocused" />
+</selector>
