+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:id="@+id/login"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginActivity">
+    tools:context=".ui.auth.LoginActivity">
 
     <ImageView
         android:id="@+id/imageViewLogin"
@@ -86,6 +86,7 @@
             android:layout_marginTop="16dp"
             android:text="Login"
             android:textSize="14sp"
+            android:backgroundTint="@color/dark_blue"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent" />
