+++ app/src/main/res/layout/activity_main.xml
@@ -17,7 +17,20 @@
            app:layout_constraintTop_toTopOf="parent">
 
 
+
        </FrameLayout>
+
+       <com.google.android.material.floatingactionbutton.FloatingActionButton
+           android:id="@+id/main_fab"
+           android:layout_width="56dp"
+           android:layout_height="56dp"
+           android:layout_marginRight="24dp"
+           android:layout_marginBottom="24dp"
+           app:layout_constraintBottom_toTopOf="@id/bottomNavigation"
+           app:layout_constraintEnd_toEndOf="parent"
+           app:backgroundTint="@color/dark_blue"
+           app:tint="@color/white"
+           app:srcCompat="@drawable/ic_plus_solid" />
 
 
        <com.google.android.material.bottomnavigation.BottomNavigationView
@@ -30,7 +43,6 @@
            app:itemTextColor="@color/color_state"
            app:layout_constraintBottom_toBottomOf="parent"
            app:layout_constraintEnd_toEndOf="parent"
-
            app:layout_constraintStart_toStartOf="parent"
            app:menu="@menu/bottom_navigation_menu" />
