+++ app/src/main/res/layout/activity_register.xml
@@ -5,7 +5,7 @@
     android:id="@+id/register"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RegisterActivity"
+    tools:context=".ui.auth.RegisterActivity"
     android:background="@color/white_but_not_white">
 
     <ImageView
@@ -106,6 +106,7 @@
             android:layout_marginTop="14dp"
             android:text="Signup"
             android:textSize="14sp"
+            android:backgroundTint="@color/dark_blue"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent" />
